panoconfig360_backend/requirements.txt
fastapi
uvicorn
pillow
numpy
requests